Cleanup namespace and move some classes to GuiBase.
